Bugfix: When the users has opened an DVC Editor and clicks somewhere else, the Editor is now closed correcty.
The problem was, that, the EVT_KILL_FOCUS was not propagated from the childs of the ObjPanel to the ObjPanel itself. Now the ObjEditor is in most cases directly a wx.Window (mostly wx.TextCtrl) so that EVT_KILL_FOCUS event dont has to be propageted somewhere else. Only in ObjEditor_Timestamp this is not possible because there are more than one child in the ObjPanel. In this case the  EVT_KILL_FOCUS is propagated to the parent panel.