Fixing a bug, for conditional constructs (IfThenElse, Switch, ...). Every dvc_item now belongs to an single EntryConstruct and is not passed to its children.
Description: In every conditional construct (IfThenElse, Switch, ...) the whole structure is shown in case of parsing error and the conditial construct has multiple dvc_items. In case every thing is okay the conditinal construct forwards everything to the subcon. But if at first everything was okay, and then an error occures, multiple EntryConstructs points to the same dvc_item and there is a recursion shown in the tree of the GUI.